Refactored Code Related to parsing of Emails
