--- app/src/main/java/com/priyank/drdelivery/shipmentDetails/data/remote/justforgit.kt
@@ -1,3 +1,0 @@
-package com.priyank.drdelivery.shipmentDetails.data.remote
-
-class justforgit
