+++ app/src/main/java/com/priyank/drdelivery/shipmentDetails/presentation/BottomNav.kt
@@ -16,12 +16,11 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
-import androidx.navigation.NavHost
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.currentBackStackEntryAsState
-import com.priyank.drdelivery.shipmentDetails.domain.BottomNavItem
+import com.priyank.drdelivery.shipmentDetails.domain.model.BottomNavItem
 
 @Composable
 fun BottomNavigationBar(
